// Neptune.Web/src/app/shared/generated/model/custom-attribute-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CustomAttributeUpsertDto { 
    CustomAttributeID?: number | null;
    TreatmentBMPID?: number | null;
    TreatmentBMPTypeID?: number | null;
    TreatmentBMPTypeCustomAttributeTypeID?: number;
    CustomAttributeTypeID?: number;
    CustomAttributeValues?: Array<string> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CustomAttributeUpsertDtoForm { 
    CustomAttributeID?: FormControl<number>;
    TreatmentBMPID?: FormControl<number>;
    TreatmentBMPTypeID?: FormControl<number>;
    TreatmentBMPTypeCustomAttributeTypeID?: FormControl<number>;
    CustomAttributeTypeID?: FormControl<number>;
    CustomAttributeValues?: FormControl<Array<string>>;
}

export class CustomAttributeUpsertDtoFormControls { 
    public static CustomAttributeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPTypeCustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeValues = (value: FormControlState<Array<string>> | Array<string> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<string>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}