// Neptune.Web/src/app/shared/generated/model/treatment-bmp-type-custom-attribute-type-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CustomAttributeTypeDto } from './custom-attribute-type-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPTypeCustomAttributeTypeDto { 
    TreatmentBMPTypeCustomAttributeTypeID?: number;
    TreatmentBMPTypeID?: number;
    CustomAttributeTypeID?: number;
    CustomAttributeType?: CustomAttributeTypeDto;
    SortOrder?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPTypeCustomAttributeTypeDtoForm { 
    TreatmentBMPTypeCustomAttributeTypeID?: FormControl<number>;
    TreatmentBMPTypeID?: FormControl<number>;
    CustomAttributeTypeID?: FormControl<number>;
    CustomAttributeType?: FormControl<CustomAttributeTypeDto>;
    SortOrder?: FormControl<number>;
}

export class TreatmentBMPTypeCustomAttributeTypeDtoFormControls { 
    public static TreatmentBMPTypeCustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TreatmentBMPTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeType = (value: FormControlState<CustomAttributeTypeDto> | CustomAttributeTypeDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<CustomAttributeTypeDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}